==> TILDA/6_Physical_measures/1_weight.Rmd <==
---
title: "Weight"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_iago.r")
```
<!-- ########################################################## --> 



# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `weight`
* Variable label: `Weight of the participant in kg`
* Variable description: `Weight should be collected in kilograms. It should be determined wheather the  variable was collected through a self reported measure or an anthropometric test`
* Domain: `Physical measures`
* Value type: `continuous`
* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

```{r lab, echo=FALSE}
ds_label <- "Weight"
ds_label_all <- "Weight of the participant in kg"
```


# Data process

## Ireland

### Wave 1 

#### Study-specific variable description


| **Name** | `weight`|
|-|-|
| **Label** |  |
| **Table name** | `TILDA-AMF-v16` |
| **Categories** | `continuous; 45 = <=45; 135 = 135+` |
| **Description** |  |



```{r assign1, echo=FALSE}
opal.assign.table.tibble(o, 'TILDA_w1_weig','TILDA.TILDA-AMF-v16',
                         variables=list('weight'), missings = TRUE)
TILDA_w1_weig <- opal.execute(o,'TILDA_w1_weig')
```

```{r local1, echo=FALSE}
vari <- TILDA_w1_weig$weight
kable(Continuous_summary(var = vari, missing_values = NA)[3], caption = ds_label)
#kable(Continuous_summary(vari, missing_values = c(45,135))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(TILDA_w1_weig, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("Height") + ylab("Frequency")
```

The variable 'weight' is categorized on the extreme values.


#### Harmonisation algorithm


The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`
* `The extreme values are preserved`

**R script:**

```{r harmo1}
TILDA_ds_w1 <- tibble(id=TILDA_w1_weig$id)
TILDA_ds_w1$weight <- car::recode(TILDA_w1_weig$weight, "NA=999")
TILDA_ds_w1$weight <- labelled(TILDA_ds_w1$weight, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript1, echo=FALSE}
vari <- TILDA_ds_w1$weight
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(TILDA_ds_w1, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("weight") + ylab("Frequency") + xlim(40,140)
```

#### Validation
```{r crosstabulation1, echo=FALSE}
```

```{r importOPAL1, echo=FALSE}
#val_labels(TILDA_ds_w1) <- NULL
#opal.assign.data(o, 'TILDA_ds_w1', TILDA_ds_w1)
#opal.symbol_import(o,'TILDA_ds_w1', project='_Athlos_Harm_Dataset')
label(TILDA_ds_w1$weight) <- ds_label_all
weight <- TILDA_ds_w1
save(weight, file = "../RData/w1/weight.RData")
```







### Wave 2 

#### Study-specific variable description



| **Name** | `ph720`|
|-|-|
| **Label** | `Weight in kilograms` |
| **Table name** | `TILDA-AMF-v23` |
| **Categories** | `continuous; 49 = Less than 50; 100 = 100+`|
| **Missings** | `-99 = Refused; -98 = Don't know; -1 = Not applicable or skipped due to routing patterns` |
| **Description** |  |




```{r assign2, echo=FALSE}
opal.assign.table.tibble(o, 'TILDA_w2_weig','TILDA.TILDA-AMF-v23',
                         variables=list('ph720'), missings = TRUE)
TILDA_w2_weig <- opal.execute(o,'TILDA_w2_weig')
```

```{r local2, echo=FALSE}
vari <- TILDA_w2_weig$ph720
kable(Continuous_summary(var = vari, missing_values = c( -99, -98, -1 ))[3], caption = ds_label)
#kable(Continuous_summary(vari, missing_values = c(-99, -98, -1 ))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-99, -98, -1 ))$summary, caption = "Summary")
ggplot(TILDA_w2_weig, aes(ph720)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency") + xlim(50,100)
```

The variable 'ph720' is categorized on the extreme values.

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `The extreme values are preserved`
* `recode -99 "Refused" to 997`
* `recode -98 "Don't know" to 998`
* `recode -1 "Not applicable or skipped due to routing patterns" to 996`
* `NA into 999`

**R script:**

```{r harmo2}
TILDA_ds_w2 <- tibble(id=TILDA_w2_weig$id)
TILDA_ds_w2$weight <- car::recode(TILDA_w2_weig$ph720, "NA=999; -99=997; -98=998; -1=996")
TILDA_ds_w2$weight <- labelled(TILDA_ds_w2$weight, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript2, echo=FALSE}
vari <- TILDA_ds_w2$weight
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(TILDA_ds_w2, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("weight") + ylab("Frequency") + xlim(49,100)
```

#### Validation
```{r crosstabulation2, echo=FALSE}
```

```{r importOPAL2, echo=FALSE}
label(TILDA_ds_w2$weight) <- ds_label_all
weight <- TILDA_ds_w2
save(weight, file = "../RData/w2/weight.RData")
#opal.assign.data(o, 'TILDA_ds_w2', TILDA_ds_w2)
#opal.symbol_import(o,'TILDA_ds_w2', project='_Athlos_Harm_Dataset')
```



# Quality estimation
Anthropometric measure.


```{r closeRsession, echo=FALSE}
opal.logout(o)
```
