==> TILDA/6_Physical_measures/4_Obesity.Rmd <==
---
title: "Obesity"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ############################ --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_iago.r")
```
<!-- ############################ --> 



# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `obesity`
* Variable label: `Presence of obesity based in Body Mass Index classification according to the WHO BMI classification (http://apps.who.int/bmi/index.jsp?introPage=intro_3.html): BMI value equal to or above 30Kg/m^2`
* Variable description: `the BMI variable should be created by dividing weight and hiegh variables by means of the following equation= weight (kg)/height(m)^2. Then, the Obesity variable should be generated according to the WHO cutoff point for BMI classification`
* Domain: `Physical measures`
* Value type: `categorical`
* Category coding:

**Code** | **Category Label**
-------- | ------------------
0        | not obese
1        | obese

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing


```{r lab, echo=FALSE}
ds_label <- "Obesity"
ds_label_all <- "Precense of obesity based in Body Mass Index classification according to the WHO BMI classification (http://apps.who.int/bmi/index.jsp?introPage=intro_3.html): BMI value equal to or above 30Kg/m^2"
```

# Data process

## Ireland

### Wave 1 

#### Study-specific variable description



| **Name** | `FRbmi`|
|-|-|
| **Label** |  |
| **Table name** | `TILDA-AMF-v16` |
| **Categories** | `continuous; 18 = <18.5; 36 = [36,40); 40 = [40,45); 45 = >=45` |
| **Description** |  |



```{r assign1, echo=FALSE}
opal.assign.table.tibble(o, 'TILDA_bmi','TILDA.TILDA-AMF-v16',
                         variables=list('FRbmi'), missings = TRUE)
TILDA_bmi <- opal.execute(o,'TILDA_bmi')
```

```{r local11, echo=FALSE}
vari <- TILDA_bmi[[2]]
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(TILDA_bmi, aes(x=TILDA_bmi[[2]])) + geom_bar(stat="count", width=0.1, fill="steelblue") + xlab("BMI") + ylab("Frequency")
```

#### Harmonisation algorithm

To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

**R script:**

```{r harmo11}
TILDA_ds_w1 <- tibble(id=TILDA_bmi$id)
TILDA_ds_w1$obesity <- car::recode(TILDA_bmi[[2]], "lo:29.9999999=0; 30:hi=1; NA=999")
TILDA_ds_w1$obesity <- labelled(TILDA_ds_w1[[2]], labels = c("not obese"=0, "obese"=1,  "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript11, echo=FALSE}
vari <- TILDA_ds_w1$obesity
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(TILDA_ds_w1, aes(x=factor(obesity))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")
```

#### Validation
```{r crosstab11, echo=FALSE}
BeforeH <- c(sum(na.omit(TILDA_bmi[[2]])<30),sum(na.omit(TILDA_bmi[[2]])>=30), sum(is.na(TILDA_bmi[[2]])))
AfterH <- table(TILDA_ds_w1[[2]], useNA = "ifany")
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("<30->0",">=30->1","NA->999") 
kable(C)
```


```{r importOPAL EBB w7, echo=FALSE}
label(TILDA_ds_w1$obesity) <- ds_label_all
obesity <- TILDA_ds_w1
save(obesity, file = "../RData/w1/obesity.RData")
```


```{r read, include=FALSE}
rm(TILDA_bmi)
```



### Wave 2

#### Study-specific variable description



| **Name** | `bmi`|
|-|-|
| **Label** | `Body Mass Index calculated as weight (kg)/height(m)^2` |
| **Table name** | `bmi` |
| **Categories** | `continuous; 18 = <18.5; 36 = [36,40); 40 = [40,45); 45 = >=45` |
| **Missings** | `991 = CAPI/interviewer error; 992 = Impute; 993 = Disable to measure; 994 = Not attempt/not done; 995 = Does not answer; 996 = Not applicable; 997 = Refuse; 998 = Do not know; 999 = Missing` |
| **Description** |  |


    
```{r assign_fi1, include=FALSE}
load("../RData/w2/bmi.RData")
TILDA_bmi <- bmi
```

```{r local2, echo=FALSE}
vari <- TILDA_bmi[[2]]
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(TILDA_bmi, aes(x=round(TILDA_bmi[[2]],3))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("BMI") + ylab("Frequency") + xlim(10,45)
```

#### Harmonisation algorithm

To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

**R script:**

```{r harmo2}
TILDA_ds_w2 <- tibble(id=TILDA_bmi$id)
TILDA_ds_w2$obesity <- car::recode(TILDA_bmi[[2]], "lo:29.9999999=0; 30:990=1; NA=999")
TILDA_ds_w2$obesity <- labelled(TILDA_ds_w2[[2]], labels = c("not obese"=0, "obese"=1,  "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript2, echo=FALSE}
vari <- TILDA_ds_w2$obesity
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(TILDA_ds_w2, aes(x=factor(obesity))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")
```

#### Validation
```{r crosstab2, echo=FALSE}
BeforeH <- c(sum(na.omit(TILDA_bmi[[2]])<30),sum(na.omit(TILDA_bmi[[2]])>=30 & na.omit(TILDA_bmi[[2]])<=990),sum(na.omit(TILDA_bmi[[2]])==996))
AfterH <- table(TILDA_ds_w2[[2]], useNA = "ifany")
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("<30->0",">=30->1","996") 
kable(C)
```


```{r importOPAL2, echo=F}
label(TILDA_ds_w2$obesity) <- ds_label_all
obesity <- TILDA_ds_w2
save(obesity, file = "../RData/w2/obesity.RData")
```

```{r read22, include=FALSE}
rm(TILDA_bmi)
```

<!--- ############################# Import to OPAL ###########################-->
```{r importOPAL, echo=FALSE}
TILDA_m_ds_w1   <- TILDA_ds_w1
TILDA_m_ds_w2   <- TILDA_ds_w2

l.hds <- list(TILDA_ds_w1 = TILDA_ds_w1 , TILDA_ds_w2 = TILDA_ds_w2)

m.hds <- list(TILDA_m_ds_w1 = TILDA_m_ds_w1 , TILDA_m_ds_w2 = TILDA_m_ds_w2)


for(i in 1:length(l.hds)) {
  label(l.hds[[i]][[2]]) <- label(m.hds[[i]][[2]]) <- ds_label # Text value assigned at the DS description, at the top.
    
  m.hds[[i]][[2]] <- car::recode(m.hds[[i]][[2]], "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
  m.hds[[i]][[2]] <- labelled(m.hds[[i]][[2]], labels = c("not obese"=0, "obese"=1, "Respond, died after this wave"=2, "Respond(proxy), died"=3, "No respond, alive"=4, "No respond, died this wave"=5, "No respond, died previous waves"=6, "No respond, dropped from sample"=7))

  nl <- names(l.hds[i])
  #opal.assign.data(o, nl , l.hds[[i]])
  #opal.symbol_import(o, nl , project='_Athlos_Harm_Dataset')  

  ml <- names(m.hds[i])
  #opal.assign.data(o, ml , m.hds[[i]]) 
  #opal.symbol_import(o, ml , project='_Athlos_Harm_Dataset')  

}
```



## Descriptive statistics of the harmonised variable accross populations and waves

Percentages of categories in each harmonised variable by cohort and wave:

**R script:**

```{r summ}
# Table HRS
t.hds <- frq(l.hds[[2]][2])[[1]][,c(1,2)] 
for (i in 1:length(l.hds)){
  t.hds[2+i] <- frq(l.hds[[i]][2])[[1]][,4] 
}
t.hds[13,] <- c("n", "sample size", 
             length(l.hds[[1]][[2]]), length(l.hds[[2]][[2]]))
names(t.hds)<-c("val", "label", "Wave1", "Wave2")
kable(t.hds)
```


# Quality estimation
Once BMI has been calculated using weight (Kg) and Height (m^2),
Obesity was calculated by using the BMI cutoff>=30kg/m2 following the WHO 
classification. Finally, the resulting BMI variable was coded as follows: 
BMI>30kg/m2 =1 (obese); BMI<30kg/m2 =0 (not obese) to generate two categories.


<!--- #### Close OPAL R Session -->
```{r closeRsession, echo=FALSE}
#opal.logout(o)
```
