==> TILDA/6_Physical_measures/7_Grip_Strength.Rmd <==
---
title: "Grip strength"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_laia.r")
```
<!-- ########################################################## --> 



# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `grip`
* Variable label: `Grip strength in kg`
* Variable description: `Grip strength`
* Domain: `Physical measures`
* Value type: `continuous`

* Comments: `According to a systematic review (Roberts et al, 2011), there are some inconsistencies in the use of maximum or mean grip strength as a summary measure which limit comparison of results between epidemiological studies. With multiple attempts, the maximum grip strength will be greater than the mean value. 
Even though for right handed people, there is a 10% stronger grip in the dominant hand, this is not true for left handed or ambidextrous people, whose grip strength are more or less equal in both hands (Roberts et al, 2011). 
In light of the above, the harmonised variable "grip" is proposed to be the highest score (to the nearest kg) of all hand grip measurements in each study, regardless of hand dominance. This is following a protocol suggested by Roberts et al (2011), based on the American Society of Hand Therapists (ASHT) protocol.`

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

```{r lab, echo=FALSE}
ds_label <- "Grip strength"
```

# Data process

## Ireland

### Wave 1 

#### Study-specific variable description


| **Name** | `GRIPtest1D`| `GRIPtest2D`| `GRIPtest1ND`| `GRIPtest2ND`|
|-|-|-|-|-|
| **Label** | `Grip Test 1 - Dominant Hand` | `Grip Test 2 - Dominant Hand` | `Grip Test 1 - Non-Dominant Hand` | `Grip Test 2 - Non-Dominant Hand` |
| **Table name** | `TILDA-AMF-v16` |`TILDA-AMF-v16` |`TILDA-AMF-v16` |`TILDA-AMF-v16` |
| **Categories** | `continuous` |`continuous` |`continuous` |`continuous` |
| **Description** |  | | | |


```{r assign1, echo=FALSE}
opal.assign.table.tibble(o, 'TILDA_gs','TILDA.TILDA-AMF-v16',
                         variables=list('GRIPtest1D', 'GRIPtest2D', 'GRIPtest1ND', 'GRIPtest2ND'), missings = TRUE)
TILDA_gs <- opal.execute(o,'TILDA_gs')
```

```{r local1, echo=FALSE}
vari <- TILDA_gs$GRIPtest1D
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(TILDA_gs, aes(GRIPtest1D)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")

vari <- TILDA_gs$GRIPtest2D
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(TILDA_gs, aes(GRIPtest2D)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")

vari <- TILDA_gs$GRIPtest1ND
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(TILDA_gs, aes(GRIPtest1ND)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")

vari <- TILDA_gs$GRIPtest2ND
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(TILDA_gs, aes(GRIPtest2ND)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```


#### Harmonisation algorithm
The harmonized variable is the higest score of all measurements and the missing values have to be recoded as follows: 

* `NA into 999`

**R script:**

```{r harmo1}
TILDA_ds_w1 <- tibble(id=TILDA_gs$id)
TILDA_ds_w1$grip <- apply(TILDA_gs[,2:5], MARGIN = 1, FUN = function(x) max(x, na.rm = TRUE))
TILDA_ds_w1$grip <- car::recode(TILDA_ds_w1$grip, "-Inf=999")
TILDA_ds_w1$grip <- as.integer(TILDA_ds_w1$grip)
TILDA_ds_w1$grip <- labelled(TILDA_ds_w1$grip, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript1, echo=FALSE}
vari <- TILDA_ds_w1$grip
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(TILDA_ds_w1, aes(grip)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100) 
```

#### Validation
```{r crosstabulation1, echo=FALSE}
```

```{r importOPAL1, echo=FALSE}
label(TILDA_ds_w1$grip) <- ds_label
grip <- TILDA_ds_w1
save(grip, file = "../RData/w1/grip.RData")
#val_labels(TILDA_ds_w1) <- NULL
#opal.assign.data(o, 'TILDA_ds_w1', TILDA_ds_w1)
#opal.symbol_import(o,'TILDA_ds_w1', project='_Athlos_Harm_Dataset')
```







### Wave 2 

#### Study-specific variable description


| **Name** | `gs005`| `gs006`| 
|-|-|-|-|-|
| **Label** | `gs005 - IWER: DOMINANT HAND MEASUREMENT. SAY: 'ONE, TWO, THREE, SQUEEZE!' Enter` | `gs006 - IWER: NON-DOMINANT HAND MEASUREMENT. SAY: 'ONE, TWO, THREE, SQUEEZE!' E` | 
| **Table name** | `TILDA-AMF-v23` |`TILDA-AMF-v23` |
| **Categories** | `continuous` |`continuous` |
| **Missing Categories** | `-98 = Unable to record` | `-98 = Unable to record` |
| | `-1 = Not applicable` |  `-1 = Not applicable` |
| **Description** |  | |


```{r assign2, echo=FALSE}
opal.assign.table.tibble(o, 'TILDA_gs','TILDA.TILDA-AMF-v23',
                         variables=list('gs005', 'gs006'), missings = TRUE)
TILDA_gs <- opal.execute(o,'TILDA_gs')
```

```{r local2, echo=FALSE}
vari <- TILDA_gs$gs005
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(TILDA_gs, aes(gs005)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100)

vari <- TILDA_gs$gs006
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(TILDA_gs, aes(gs006)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") #+ xlim(-1,100)
```


#### Harmonisation algorithm
The harmonized variable is the higest score of all measurements and the missing values have to be recoded as follows: 

* `-1 into 996`
* `-98 into 993`
* `NA into 999`

**R script:**

```{r harmo2}
TILDA_ds_w2 <- tibble(id=TILDA_gs$id)
TILDA_ds_w2$grip <- apply(TILDA_gs[,2:3], MARGIN = 1, FUN = function(x) max(x, na.rm = TRUE))
TILDA_ds_w2$grip <- car::recode(TILDA_ds_w2$grip, "-Inf=999; -1=996")
TILDA_ds_w2$grip <- as.integer(TILDA_ds_w2$grip)
TILDA_ds_w2$grip <- labelled(TILDA_ds_w2$grip, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript2, echo=FALSE}
vari <- TILDA_ds_w2$grip
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(TILDA_ds_w2, aes(grip)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100) 
```

#### Validation
```{r crosstabulation2, echo=FALSE}
```

```{r importOPAL2, echo=FALSE}
label(TILDA_ds_w2$grip) <- ds_label
grip <- TILDA_ds_w2
save(grip, file = "../RData/w2/grip.RData")
#val_labels(TILDA_ds_w2) <- NULL
#opal.assign.data(o, 'TILDA_ds_w2', TILDA_ds_w2)
#opal.symbol_import(o,'TILDA_ds_w2', project='_Athlos_Harm_Dataset')
```


# Quality estimation
No coments.  


```{r closeRsession, echo=FALSE} 
opal.logout(o)
```
