==> TILDA/6_Physical_measures/10_Systolic.Rmd <==
---
title: "Systolic blood pressure"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---


```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_iago.r")
```

# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `sbp`
* Variable label: `Systolic blood pressure of participant at each wave`
* Variable description: `Means of all sistolic blood pressure measures`
* Domain: `Physical measures`
* Value type: `continuous`
* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing



```{r lab, echo=FALSE}
ds_label <- "Systolic blood pressure"
ds_label_all <- "Systolic blood pressure of participant at each wave"
```

# Data process

## Ireland

### Wave 1 

#### Study-specific variable description


| **Name** | `BPseatedsystolic1`|`BPseatedsystolic2`|`BPseatedsystolicmean`|
|-|-|-|-|
| **Label** |  |  | `Seated systolic blood pressure (mean)` |
| **Table name** | `TILDA-AMF-v16` |`TILDA-AMF-v16` |`TILDA-AMF-v16` |
| **Categories** | `continuous` |`continuous` |`continuous` |
| **Description** |  |  |  |


```{r assign1, echo=FALSE}
opal.assign.table.tibble(o, 'TILDA_w1_sbp','TILDA.TILDA-AMF-v16',
                         variables=list('BPseatedsystolic1', 'BPseatedsystolic2', 'BPseatedsystolicmean'), missings = TRUE)
TILDA_w1_sbp <- opal.execute(o,'TILDA_w1_sbp')
```

```{r local1, echo=FALSE}
vari <- TILDA_w1_sbp$BPseatedsystolic1
kable(Continuous_summary(vari, missing_values = NA)[3], caption = "BPseatedsystolic1") 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(TILDA_w1_sbp, aes(BPseatedsystolic1)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

vari <- TILDA_w1_sbp$BPseatedsystolic2
kable(Continuous_summary(vari, missing_values = NA)[3], caption = "BPseatedsystolic2") 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(TILDA_w1_sbp, aes(BPseatedsystolic2)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

vari <- TILDA_w1_sbp$BPseatedsystolicmean
kable(Continuous_summary(vari, missing_values = NA)[3], caption = "BPseatedsystolicmean") 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(TILDA_w1_sbp, aes(BPseatedsystolicmean)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")
```

Since the variable 'BPseatedsystolicmean' is the mean beetwen the variables 'BPseatedsystolic1' and 'BPseatedsystolic2', the variable is going to be harmonized from 'BPseatedsystolicmean'. 

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`


**R script:**

```{r harmo1}
TILDA_ds_w1 <- tibble(id=TILDA_w1_sbp$id)
TILDA_ds_w1$sbp <- car::recode(TILDA_w1_sbp$BPseatedsystolicmean, "NA=999")
TILDA_ds_w1$sbp <- labelled(TILDA_ds_w1$sbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript1, echo=FALSE}
vari <- TILDA_ds_w1$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(TILDA_ds_w1, aes(round(sbp,3))) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency") + xlim(70,230) 
```

#### Validation
```{r crosstabulation1, echo=FALSE}
TILDA_w1_sbp$mean_sbp <- apply(X = TILDA_w1_sbp[,c(2,3)], MARGIN = 1, FUN = mean)
TILDA_w1_sbp$proff <- TILDA_w1_sbp$mean_sbp - TILDA_w1_sbp$BPseatedsystolicmean
pander(summary(TILDA_w1_sbp$proff), caption = "Summary of mean(BPseatedsystolic1,BPseatedsystolic2) - BPseatedsystolicmean")
```

```{r importOPAL1, echo=FALSE}
#opal.assign.data(o, 'TILDA_ds_w1', TILDA_ds_w1)
#opal.symbol_import(o,'TILDA_ds_w1', project='_Athlos_Harm_Dataset')

#label(TILDA_ds_w1) <- ds_label_all
label(TILDA_ds_w1$sbp) <- ds_label_all
sbp <- TILDA_ds_w1
save(sbp, file = "../RData/w1/sbp.RData")
```







### Wave 2 

There is no variable. 

# Quality estimation
Systolic bp was calculated as the mean value of the two available variables.


```{r closeRsession, echo=FALSE} 
opal.logout(o)
```
